backend/orchestrator/requirements.txt
fastapi>=0.116.0
uvicorn[standard]>=0.35.0
PyJWT>=2.10.1
pydantic>=2.11.0
crewai>=0.161.0
